LPS-67544 Move plugins and tasks
